add tests

+++ x-pack/plugin/transform/qa/single-node-tests/src/javaRestTest/java/org/elasticsearch/xpack/transform/integration/continuous/DateHistogramGroupByOtherTimeFieldIT.java
@@ -198,8 +198,8 @@ public class DateHistogramGroupByOtherTimeFieldIT extends ContinuousTestCase {
             );
             assertThat(
                 "Doc count did not match, source: " + source + ", expected: " + bucket.get("count") + ", iteration: " + iteration,
-                XContentMapValues.extractValue("count", source),
-                equalTo(Double.valueOf(((Long) bucket.get("count"))))
+                ((Integer) XContentMapValues.extractValue("count", source)).longValue(),
+                equalTo(bucket.get("count"))
             );
             assertThat(
                 "Term did not match, source: " + source + ", expected: " + bucket.get("event") + ", iteration: " + iteration,
@@ -217,9 +217,11 @@ public class DateHistogramGroupByOtherTimeFieldIT extends ContinuousTestCase {
                     + iteration,
                 // we use a fixed_interval of `1s`, the transform runs every `1s`, a bucket might be recalculated at the next run
                 // but should NOT be recalculated for the 2nd/3rd/... run
-                Double.valueOf((Integer) XContentMapValues.extractValue(INGEST_RUN_FIELD, source)) - (Double) XContentMapValues
-                    .extractValue(MAX_RUN_FIELD, source),
-                is(lessThanOrEqualTo(1.0))
+                (Integer) XContentMapValues.extractValue(INGEST_RUN_FIELD, source) - (Integer) XContentMapValues.extractValue(
+                    MAX_RUN_FIELD,
+                    source
+                ),
+                is(lessThanOrEqualTo(1))
             );
         }
         assertFalse(sourceIterator.hasNext());
